services/orchestrator/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
redis==5.0.4
psycopg[binary,pool]==3.2.9
httpx==0.27.0
websockets==12.0
ruff==0.5.1
pytest==8.2.1
openai>=1.12.0
anthropic>=0.16.0
google-generative-ai>=0.3.1
